pyproject.toml
[project]
name = "LifeManager"
version = "1.0.7"
description = "A simple Python package"
authors = [{name = "Raad", email = "alirezzaraadco@gmail.com"}]
requires-python = ">=3.12"
dependencies = [
  "matplotlib>=3.10.1",
  "sqlalchemy>=2.0.40",
  "pandas>=2.2.3",
  "python-dotenv>=1.1.0",
  "psycopg2-binary>=2.9.10",
  "openpyxl>=3.1.5",
  "aiogram==3.20.0.post0",
  "streamlit>=1.46.0"
]
license = "MIT"
readme = "README.md"

classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent"
]


[project.urls]
Homepage = "https://github.com/yourusername/LifeManager"
Documentation = "https://github.com/yourusername/LifeManager/README.md"
Repository = "https://github.com/yourusername/LifeManager.git"
Issues = "https://github.com/yourusername/LifeManager/issues"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["LifeManager", "LifeManager.telegram", "LifeManager.LocalUI"]

[tool.setuptools.package-data]
"LifeManager" = ["*.pyi", "**/*.pyi","py.typed"]